requirements.txt
openai>=1.14.0
pydantic>=2.6.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
